feat: form footer & categories select
